Normalize /core/Path, and apply MrDoob eslint style
